Rename dir - 2 acceleration_management
