+++ src/plugins/data_source_management/public/components/direct_query_data_sources_components/acceleration_management/acceleration_action_overlay.tsx
