+++ src/plugins/data_source_management/public/components/direct_query_data_sources_components/acceleration_management/flyout_modules/acceleration_details_tab.test.tsx
