+++ src/plugins/data_source_management/public/components/direct_query_data_sources_components/acceleration_management/flyout_modules/accelerations_schema_tab.tsx
