+++ src/plugins/data_source_management/public/components/direct_query_data_sources_components/direct_query_connection_detail/direct_query_connection_detail.test.tsx
@@ -46,7 +46,7 @@ jest.mock('../../../../framework/catalog_cache/cache_loader', () => ({
   })),
 }));
 
-jest.mock('../direct_query_acceleration_management/acceleration_table', () => ({
+jest.mock('../acceleration_management/acceleration_table', () => ({
   AccelerationTable: () => <div>Acceleration Table</div>,
 }));
